requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg[binary]>=3.1
dj-database-url>=2.1
python-dotenv>=1.0
